Moved the about card
